+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 0.6.1 - 2023-06-19
+
 ### Fixed
 
-- DRIFT-613: Use new Client.walk method to improve performance, [PR-12](https://github.com/panda-official/DriftCLI/pull/12)
+- DRIFT-613: Use new `Client.walk` method to improve performance, [PR-12](https://github.com/panda-official/DriftCLI/pull/12)
 
 ## 0.6.0 - 2023-04-12